Parsing finished
Managed to figure out the parsing. If this playground is run it will print each class line by line and will include all the information that is listed on the site. It takes time to parse through the site character by character so I suggest we save the last 10 semesters locally so they will be displayed quicker. Should be easy to build a GUI around this.